--- a/package.json
+++ b/package.json
@@ -4,20 +4,14 @@
   "description": "An easier way to deal with github notifications",
   "main": "main.js",
   "dependencies": {
-    "electron": "^0.4.1",
-    "electron-prebuilt-compile": "^1.2.5",
     "react": "^15.2.1",
     "react-dom": "^15.2.1"
   },
   "devDependencies": {
-    "electron-prebuilt": "^1.2.6",
+    "electron-prebuilt-compile": "^1.2.5",
     "eslint": "^3.0.1",
     "eslint-config-airbnb": "^9.0.1",
-    "eslint-plugin-import": "^1.10.2",
-    "eslint-plugin-jsx-a11y": "^2.0.0",
-    "eslint-plugin-react": "^5.2.2",
     "mocha": "^2.5.3",
-    "spectron": "^3.2.6"
   },
   "scripts": {
     "start": "./node_modules/.bin/electron .",