--- a/package.json
+++ b/package.json
@@ -8,14 +8,17 @@
     "react-dom": "^15.2.1"
   },
   "devDependencies": {
+    "electron-mocha": "^2.3.1",
     "electron-prebuilt-compile": "^1.2.5",
     "eslint": "^3.0.1",
     "eslint-config-airbnb": "^9.0.1",
+    "jsdom": "^9.4.1",
     "mocha": "^2.5.3",
+    "react-addons-test-utils": "^15.2.1"
   },
   "scripts": {
     "start": "./node_modules/.bin/electron .",
-    "test": "mocha"
+    "test": "./node_modules/.bin/electron-mocha --recursive --compilers js:babel-register test"
   },
   "repository": {
     "type": "git",